Add 2 new ekf files
These aren't tested. Option 1 is supposed to take the linear x and y as control inputs in the prediction phase (I think it assumes they are low noise measurements) to directly predict the next velocity and position. Option 2 treats the acceleration data as noisy measurements similar to the distances measured by the 3 beacons.